+++ packages/cactus-cmd-api-server/src/main/typescript/proto.d.ts
@@ -1,0 +1,5 @@
+
+declare module '*.proto' {
+  const fileContent: string;
+  export default fileContent;
+}