+++ packages/cactus-plugin-ledger-connector-besu/src/test/typescript/integration/plugin-ledger-connector-besu/deploy-contract/get-balance.test.ts
@@ -59,6 +59,7 @@ test("can get balance of an account", async (t: Test) => {
     instanceId: uuidv4(),
     pluginRegistry: new PluginRegistry({ plugins: [keychainPlugin] }),
   });
+  await connector.onPluginInit();
 
   const req: GetBalanceV1Request = { address: firstHighNetWorthAccount };
   const currentBalance = await connector.getBalance(req);