+++ packages/cactus-plugin-ledger-connector-iroha/src/main/typescript/prometheus-exporter/data-fetcher.ts
@@ -1,0 +1,10 @@
+import { Transactions } from "./response.type";
+
+import { totalTxCount, K_CACTUS_IROHA_TOTAL_TX_COUNT } from "./metrics";
+
+export async function collectMetrics(
+  transactions: Transactions,
+): Promise<void> {
+  transactions.counter++;
+  totalTxCount.labels(K_CACTUS_IROHA_TOTAL_TX_COUNT).set(transactions.counter);
+}