+++ tools/ci.sh
@@ -95,29 +95,20 @@ function mainTask()
 
   yarn run configure
 
-  yarn run custom-checks
-  node ./tools/validate-bundle-names.js 
+  # Obtains the major NodeJS version such as "12" from "v12.14.1"
+  # We only run the custom checks above v12 because the globby dependency's
+  # latest version is forcing us to use Ecmascript Modules which do not work
+  # on NodeJS 12 even with the additional flags passed in.
+  nodejs_version=`node --version | awk -v range=1 '{print substr($0,range+1,2)}'`
+  if [ "$nodejs_version" -gt "12" ]; then
+    echo "$(date +%FT%T%z) [CI] NodeJS is newer than v12, running custom checks..."
+    yarn run custom-checks
+  fi
 
-  # Tests are still flaky (on weak hardware such as the CI env) despite our best
-  # efforts so here comes the mighty hammer of brute force. 3 times the charm...
-  {
-    dumpDiskUsageInfo
-    yarn run test:all -- --bail && echo "$(date +%FT%T%z) [CI] First (#1) try of tests succeeded OK."
-  } ||
-  {
-    dumpDiskUsageInfo
-    echo "$(date +%FT%T%z) [CI] First (#1) try of tests failed starting second try now..."
-    yarn run test:all -- --bail && echo "$(date +%FT%T%z) [CI] Second (#2) try of tests succeeded OK."
+  node ./tools/validate-bundle-names.js
 
-  } ||
-  {
-    dumpDiskUsageInfo
-    # If the third try fails then the execution will reach the last echo and the exit 1 statement
-    # ensuring that the script crashes if 3 out of 3 test runs have failed.
-    echo "$(date +%FT%T%z) [CI] Second (#2) try of tests failed starting third and last try now..."
-    yarn run test:all -- --bail && echo "$(date +%FT%T%z) [CI] Third (#3) try of tests succeeded OK." || \
-      echo "$(date +%FT%T%z) [CI] Third (#3) try of tests failed so giving up at this point" ; exit 1
-  }
+  dumpDiskUsageInfo
+  yarn run test:all -- --bail
 
   dumpDiskUsageInfo